app/requirements.txt
Django==3.1.7
psycopg2-binary==2.8.5
gunicorn==20.0.4
djangorestframework
